requirements.txt
asgiref==3.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
dj-database-url==3.0.1
Django==4.2.23
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
geographiclib==2.0
geopy==2.4.1
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
razorpay==1.4.2
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
